- restructure the back apps - fix activities and attends state update - update frontend root to new backend structure
