+++ .gitignore
@@ -189,4 +189,4 @@ data
 # Backend
 apps/backend/Pipfile
 apps/backend/Pipfile.lock
-apps/backend/api/migrations/0001_initial.py
+apps/backend/authentification/migrations/0001_initial.py
