+++ apps/backend/activities/urls.py
@@ -1,6 +1,6 @@
+from activities.views import ActivityViewSet, SiteViewSet, AttendViewSet, RoomViewSet, RegisterToActivityView
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from activities.views import ActivityViewSet, SiteViewSet, AttendViewSet, RoomViewSet
 
 router = DefaultRouter()
 
@@ -8,6 +8,7 @@ router.register(r'sites', SiteViewSet)
 router.register(r'attends', AttendViewSet)
 router.register(r'rooms', RoomViewSet)
 router.register(r'activity', ActivityViewSet)
+router.register(r'register_activity', RegisterToActivityView)
 
 urlpatterns = [
     path('', include(router.urls)),
