+++ apps/backend/authentification/serializers.py
@@ -1,0 +1,46 @@
+from database.models import User
+from django.contrib.auth import authenticate
+from rest_framework import serializers
+from rest_framework.exceptions import ValidationError
+
+
+class UserRegistrationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = '__all__'
+
+    def create(self, clean_data):
+        user_obj = User.objects.create_user(**clean_data)
+        return user_obj
+
+
+class UserLoginSerializer(serializers.Serializer):
+    username = serializers.CharField()
+    password = serializers.CharField(write_only=True)
+
+    class Meta:
+        model = User
+        fields = "__all__"
+
+    def check_user(self, clean_data):
+        user = authenticate(**clean_data)
+        if not user:
+            raise ValidationError("User not found")
+        return user
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = "__all__"
+
+
+class UpdateUserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('id', 'username', 'first_name', 'last_name', 'email', 'noma')
+
+    def validate_noma(self, value):
+        if len(value) != 8 and len(value) != 0:
+            raise ValidationError("Le noma doit contenir 8 caractères")
+        return value
