+++ apps/backend/authentification/urls.py
@@ -7,14 +7,7 @@ from .views import UpdateProfileView
 
 router = routers.DefaultRouter()
 
-router.register(r'register_activity',
-                views.RegisterToActivityView, "register_activity")
-router.register(r'activity', views.ActivityViewSet, "activity")
-router.register(r'attends', views.AttendViewSet, "attends")
 router.register(r'user_profile', UpdateProfileView, "user_profile")
-router.register(r'room', views.RoomViewSet, "room")
-router.register(r'site', views.SiteViewSet, "site")
-
 
 urlpatterns = [
     path('register/', views.UserRegisterView.as_view(), name='register'),
