+++ apps/frontend/go4success/app/_layout.tsx
@@ -1,6 +1,6 @@
 import { AuthProvider } from "@/context/auth";
 import { Slot } from "expo-router";
-import { RootSiblingParent } from "react-native-root-siblings";
+
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import Toast from "react-native-toast-message";
 import { toastConfig } from "@/components/ToastConfig";
@@ -12,10 +12,8 @@ export default function Layout() {
     return (
         <QueryClientProvider client={queryClient}>
             <AuthProvider>
-                <RootSiblingParent>
-                    <Slot />
-                    <Toast visibilityTime={2000} config={toastConfig} />
-                </RootSiblingParent>
+                <Slot />
+                <Toast visibilityTime={2000} config={toastConfig} />
             </AuthProvider>
         </QueryClientProvider>
     );
