+++ apps/frontend/go4success/assets/images.d.ts
@@ -2,3 +2,7 @@ declare module "*.png" {
     const value: any;
     export = value;
 }
+declare module "*.jpg" {
+    const value: any;
+    export = value;
+}
